Make step indicator component
