+++ libs/design-system/step-indicator/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/step-indicator.component';
