+++ libs/design-system/step-indicator/src/lib/step-indicator.component.html
@@ -1,0 +1,1 @@
+<span class="step-number">{{ value() }}</span>
