+++ libs/design-system/step-indicator/src/lib/step-indicator.component.scss
@@ -1,0 +1,14 @@
+:host {
+  .step-number {
+    font: var(--sys-label-large);
+    letter-spacing: var(--sys-label-large-tracking);
+    background-color: var(--sys-inverse-surface);
+    color: var(--sys-on-primary);
+    display: flex;
+    width: 2rem;
+    height: 2rem;
+    border-radius: 50%;
+    justify-content: center;
+    align-items: center;
+  }
+}
