+++ libs/design-system/step-indicator/src/lib/step-indicator.component.ts
@@ -1,0 +1,18 @@
+import { CommonModule } from '@angular/common';
+import { ChangeDetectionStrategy, Component, input } from '@angular/core';
+
+/**
+ * Step indicator for module components that have multiple steps
+ */
+@Component({
+  selector: 'hra-step-indicator',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './step-indicator.component.html',
+  styleUrl: './step-indicator.component.scss',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+})
+export class StepIndicatorComponent {
+  /** Step value */
+  value = input.required<number>();
+}
