+++ tsconfig.base.json
@@ -113,6 +113,9 @@
       "@hra-ui/design-system/social-media-button": [
         "libs/design-system/social-media-button/src/index.ts"
       ],
+      "@hra-ui/design-system/step-indicator": [
+        "libs/design-system/step-indicator/src/index.ts"
+      ],
       "@hra-ui/design-system/table": [
         "libs/design-system/table/src/index.ts"
       ],
